[syft] move abstract node into type
